Cleaned up project structure.
