+++ src/components/signup/Signup.js
@@ -1,22 +1,24 @@
 "use client";
 
-import {useState} from "react";
-import {useRouter} from "next/navigation";
-import {Button, TextField, Typography, Box} from "@mui/material";
-import {useEffect} from "react";
+import { useState, useEffect } from "react";
+import { useRouter } from "next/navigation";
+import { Button, TextField, Typography, Box, Alert } from "@mui/material";
 
 export default function Signup() {
     const router = useRouter();
 
     const [formData, setFormData] = useState({
-        name: '',
-        email: '',
-        major: '',
-        gradYear: '',
-        preferredLanguage: '',
-        password: '',
-        confirmPassword: '',
-    })
+        name: "",
+        email: "",
+        major: "",
+        gradYear: "",
+        preferredLanguage: "",
+        password: "",
+        confirmPassword: "",
+    });
+
+    const [errorMessage, setErrorMessage] = useState("");
+    const [successMessage, setSuccessMessage] = useState("");
 
     useEffect(() => {
         document.body.style.overflowY = "hidden";
@@ -36,43 +38,59 @@ export default function Signup() {
 
         if (!formData.major) errors.major = "Major is required";
         if (!formData.gradYear) errors.gradYear = "Graduation Year is required";
+        if (!formData.preferredLanguage) errors.preferredLanguage = "Preferred Language is required";
         if (!formData.password) errors.password = "Password is required";
         if (!formData.confirmPassword) errors.confirmPassword = "Confirm Password is required";
-        if (formData.password !== formData.confirmPassword) errors.confirmPassword = "Passwords do not match";
+        if (formData.password !== formData.confirmPassword)
+            errors.confirmPassword = "Passwords do not match";
+
         return errors;
-    }
+    };
 
     const handleBackToLoginClick = () => {
         router.push("/login");
     };
 
-    const handleSubmit = async() => {
-        if(Object.keys(validateForm()).length > 0) {
-            console.log("Form is invalid");
+    const handleSubmit = async () => {
+        setErrorMessage("");
+        setSuccessMessage("");
+
+        const errors = validateForm();
+        if (Object.keys(errors).length > 0) {
+            setErrorMessage(Object.values(errors)[0]); // show first validation error
             return;
         }
 
-        console.log(JSON.stringify(formData));
+        try {
+            const response = await fetch("/api/users", {
+                method: "POST",
+                headers: { "Content-Type": "application/json" },
+                body: JSON.stringify(formData),
+            });
 
-        const response = await fetch('/api/user', {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify(formData),
-        })
+            const data = await response.json();
+            console.log("response:", data);
 
-        console.log("response status: ", response.status);
-        console.log("response data: ", await response.json());
-    }
+            if (!response.ok) {
+                setErrorMessage(data.error || "Error creating account");
+                return;
+            }
+
+            setSuccessMessage("Account created successfully! Redirecting...");
+            setTimeout(() => router.push("/login"), 1500);
+        } catch (error) {
+            console.error("Signup error:", error);
+            setErrorMessage("Something went wrong. Please try again.");
+        }
+    };
 
     const handleChange = (e) => {
-        const {name, value} = e.target;
+        const { name, value } = e.target;
         setFormData((prev) => ({
             ...prev,
             [name]: value,
         }));
-    }
+    };
 
     return (
         <Box display="flex" height="100vh" overflow="hidden">
@@ -88,7 +106,7 @@ export default function Signup() {
                     px: 8,
                 }}
             >
-                <Typography variant="h2" sx={{fontWeight: "normal", lineHeight: 1.2, mt: 10}}>
+                <Typography variant="h2" sx={{ fontWeight: "normal", lineHeight: 1.2, mt: 10 }}>
                     Welcome to
                 </Typography>
 
@@ -96,10 +114,10 @@ export default function Signup() {
                     component="img"
                     src="/newlogotransparent.png"
                     alt="Studi logo"
-                    sx={{width: 300, height: 95, ml: -2.5}}
+                    sx={{ width: 300, height: 95, ml: -2.5 }}
                 />
 
-                <Typography variant="h5" sx={{lineHeight: 1.3, mt: 3}}>
+                <Typography variant="h5" sx={{ lineHeight: 1.3, mt: 3 }}>
                     Meet people who have similar study needs, boosting productivity, while
                     forming new connections.
                 </Typography>
@@ -107,7 +125,7 @@ export default function Signup() {
                     component="img"
                     src="/lilstudybuds.png"
                     alt="graphic of two dogs studying"
-                    sx={{width: 450, height: "auto", mt: 5, mb: -40}}
+                    sx={{ width: 450, height: "auto", mt: 5, mb: -40 }}
                 />
             </Box>
 
@@ -123,72 +141,82 @@ export default function Signup() {
                     mt: -5,
                 }}
             >
-                <Typography variant="h3" sx={{fontWeight: "normal", mb: 4, textAlign: "center"}}>
+                <Typography variant="h3" sx={{ fontWeight: "normal", mb: 4, textAlign: "center" }}>
                     Create An Account
                 </Typography>
 
-                <TextField label="First and Last Name"
-                           name="name"
-                           fullWidth={false}
-                           sx={{width: "80%", mb: 2}}
-                           value={formData.name}
-                           onChange={handleChange}
-                           required
+                {errorMessage && (
+                    <Alert severity="error" sx={{ width: "80%", mb: 2 }}>
+                        {errorMessage}
+                    </Alert>
+                )}
+                {successMessage && (
+                    <Alert severity="success" sx={{ width: "80%", mb: 2 }}>
+                        {successMessage}
+                    </Alert>
+                )}
+
+                <TextField
+                    label="First and Last Name"
+                    name="name"
+                    sx={{ width: "80%", mb: 2 }}
+                    value={formData.name}
+                    onChange={handleChange}
+                    required
                 />
-                <TextField label="Email"
-                           name="email"
-                           type="email"
-                           fullWidth={false}
-                           sx={{width: "80%", mb: 2}}
-                           value={formData.email}
-                           onChange={handleChange}
-                           required
+                <TextField
+                    label="Email"
+                    name="email"
+                    type="email"
+                    sx={{ width: "80%", mb: 2 }}
+                    value={formData.email}
+                    onChange={handleChange}
+                    required
                 />
                 <TextField
                     label="Major"
                     name="major"
-                    fullWidth={false}
-                    sx={{width: "80%", mb: 2}}
+                    sx={{ width: "80%", mb: 2 }}
                     value={formData.major}
                     onChange={handleChange}
                     required
                 />
-                <TextField label="Graduation Year"
-                           name="gradYear"
-                           fullWidth={false}
-                           sx={{width: "80%", mb: 2}}
-                           value={formData.gradYear}
-                           onChange={handleChange}
-                           required
+                <TextField
+                    label="Graduation Year"
+                    name="gradYear"
+                    sx={{ width: "80%", mb: 2 }}
+                    value={formData.gradYear}
+                    onChange={handleChange}
+                    required
                 />
-                <TextField label="Preferred Language"
-                           name="preferredLanguage"
-                           fullWidth={false}
-                           sx={{width: "80%", mb: 2}}
-                           value={formData.preferredLanguage}
-                           onChange={handleChange}
-                           required
+                <TextField
+                    label="Preferred Language"
+                    name="preferredLanguage"
+                    sx={{ width: "80%", mb: 2 }}
+                    value={formData.preferredLanguage}
+                    onChange={handleChange}
+                    required
                 />
-                <TextField label="Password"
-                           name="password"
-                           type="password"
-                           fullWidth={false}
-                           sx={{width: "80%", mb: 3}}
-                           value={formData.password}
-                           onChange={handleChange}
-                           required
+                <TextField
+                    label="Password"
+                    name="password"
+                    type="password"
+                    sx={{ width: "80%", mb: 3 }}
+                    value={formData.password}
+                    onChange={handleChange}
+                    required
                 />
-                <TextField label="Confirm Password"
-                           name="confirmPassword"
-                           type="password"
-                           value={formData.confirmPassword}
-                           onChange={handleChange}
-                           sx={{width: "80%", mb: 3}}
-                           required
+                <TextField
+                    label="Confirm Password"
+                    name="confirmPassword"
+                    type="password"
+                    sx={{ width: "80%", mb: 3 }}
+                    value={formData.confirmPassword}
+                    onChange={handleChange}
+                    required
                 />
                 <Button
                     variant="contained"
-                    fullWidth={false}
                     sx={{
                         width: "80%",
                         height: 50,
@@ -207,14 +235,13 @@ export default function Signup() {
 
                 <Button
                     variant="contained"
-                    fullWidth={false}
                     sx={{
                         width: "80%",
                         height: 50,
                         fontWeight: "bold",
                         color: "#000",
                         backgroundColor: "#8ee5f4",
-                        "&:hover": {backgroundColor: "#6fd2e0"},
+                        "&:hover": { backgroundColor: "#6fd2e0" },
                         mt: 2,
                     }}
                     onClick={handleBackToLoginClick}
